Add initial react query
