+++ package.json
@@ -14,6 +14,7 @@
     "@fontsource/roboto": "^4.5.8",
     "@mui/icons-material": "^5.10.9",
     "@mui/material": "^5.10.9",
+    "@tanstack/react-query": "^4.13.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
